+++ src/database-server/sqlitewrapper.h
@@ -40,7 +40,7 @@
 
 #include <string>
 
-#include "src/database-server/IDataModel.h"
+#include "src/database-server/idatamodel.h"
 
 class sqlite3;
 class SQLiteWrapper : public IDataModel {
